Move some scripts, add vision class, add bad temp art.
