+++ ProjectStealth/Assets/ArtAssets/Blood Material.mat.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 9799df21c323ffe43958733369f584bc
+timeCreated: 1512247656
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
