--- ProjectStealth/Assets/Libs.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 34dc77a0614594b44b1dd7190120bf47
-folderAsset: yes
-timeCreated: 1450647131
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
