--- ProjectStealth/Assets/Libs/Physics.cs
@@ -1,18 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-
-public class Physics
-{
-
-	// Use this for initialization
-	void Start ()
-    {
-	
-	}
-	
-	// Update is called once per frame
-	void Update ()
-    {
-	
-	}
-}
