+++ ProjectStealth/Assets/Scripts/EnemyVisionField.cs.meta
@@ -1,0 +1,13 @@
+fileFormatVersion: 2
+guid: 0ffa374e26e60054f87407af233b6977
+timeCreated: 1510338128
+licenseType: Free
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
