+++ ProjectStealth/Assets/Scripts/Libs/Physics.cs
@@ -1,0 +1,18 @@
+﻿using UnityEngine;
+using System.Collections;
+
+public class Physics
+{
+
+	// Use this for initialization
+	void Start ()
+    {
+	
+	}
+	
+	// Update is called once per frame
+	void Update ()
+    {
+	
+	}
+}
